Handle NullPointerException on main activity start.
If location services is off on app start, app prompts user to turn location services on. When user selects yes and turns location services on then returns to app, network status shows as active but is not. When attempting to get the current address on creation of main activity, a NullPointerException is thrown. This is now caught and an approximate address is not entered into the location field.
It might be possible to wait and try again, but for now, let the user press the crosshair button to get the approximate address.