Dialog displays add'l info when click title/stars in results
I made a dialog that displays some additional information about a restaurant when you click its title or stars in the results page. I'll go ahead and push it, but after finishing the dialog I realized there's really only 3 additional pieces of info we might want to utilize. And that would be distance, categories (e.g. Mexican) and the phone number.
Categories and distance could easily go under the title, and if we wanted to give users the ability to call the restaurant, maybe we could put another icon up in the title bar